+++ testcases/kernel/syscalls/ftruncate/ftruncate04.c
@@ -188,7 +188,7 @@ void doparent()
 	if (sb.st_size != recstart + (RECLEN / 2)) {
 		tst_resm(TFAIL, "unexpected ftruncate failure case 4");
 		tst_resm(TFAIL, "expected size of %d, got size of %"PRId64,
-			 recstart + (RECLEN / 2), sb.st_size);
+			 recstart + (RECLEN / 2), (int64_t)sb.st_size);
 		local_flag = FAILED;
 		cleanup();
 	}
@@ -212,7 +212,7 @@ void doparent()
 	if (sb.st_size != RECLEN) {
 		tst_resm(TFAIL, "unexpected ftruncate failure case 5");
 		tst_resm(TFAIL, "expected size of %d, got size of %"PRId64,
-			 RECLEN, sb.st_size);
+			 RECLEN, (int64_t)sb.st_size);
 		local_flag = FAILED;
 		cleanup();
 	}
@@ -231,7 +231,7 @@ void doparent()
 	if (sb.st_size != (2 * len)) {
 		tst_resm(TFAIL, "unexpected ftruncate failure case 6");
 		tst_resm(TFAIL, "expected size of %d, got size of %"PRId64,
-			 (2 * len), sb.st_size);
+			 (2 * len), (int64_t)sb.st_size);
 		local_flag = FAILED;
 		cleanup();
 	}
